+++ validate/josev2/jose_test.go
@@ -1,4 +1,4 @@
-package jose_v2
+package josev2
 
 import (
 	"errors"
